requirements.txt
streamlit
pypdf
langchain
langchain-google-genai==0.0.11
google.generativeai
python-dotenv
faiss-cpu
langchain-community
beautifulsoup4
